Refactoring of Message and Node class to prepare for Flooding
Message now keeps a list of current nodes that it is contained at
(Flooding will make messages be at multiple nodes). Message also now has
a previous node to start tracking a path.

THE HOOD OF NODES IS NOW A LIST OF NODES (not edges anymore). Makes code
cleaner and will make writting future algorithms much easier.

(HAVE NOT HEAVILY TESTED CHANGES EFFECT ON THE VIEW, BUT IT WORKS IT A 
BASIC DEFAULT NETWORK)